- Change text colors - Move out some components from chatdetail - Change default character
