+++ web/pages/Chat/ChatDetail.tsx
@@ -2,15 +2,10 @@ import { A, useNavigate, useParams } from '@solidjs/router'
 import {
   ArrowDownLeft,
   ArrowUpRight,
-  Book,
   ChevronLeft,
   ChevronRight,
-  Download,
   MailPlus,
   Menu,
-  Palette,
-  Settings,
-  Sliders,
   X,
 } from 'lucide-solid'
 import ChatExport from './ChatExport'
@@ -21,11 +16,8 @@ import Button from '../../shared/Button'
 import IsVisible from '../../shared/IsVisible'
 import Modal from '../../shared/Modal'
 import TextInput from '../../shared/TextInput'
-import { Toggle } from '../../shared/Toggle'
 import { getRootRgb, getStrictForm } from '../../shared/util'
 import { chatStore, settingStore, userStore } from '../../store'
-import { AVATAR_SIZES, AvatarCornerRadius } from '../../store/user'
-import * as UserStore from '../../store/user'
 import { msgStore } from '../../store'
 import { ChatGenSettingsModal } from './ChatGenSettings'
 import ChatSettingsModal from './ChatSettings'
@@ -37,10 +29,10 @@ import DeleteMsgModal from './DeleteMsgModal'
 import './chat-detail.css'
 import { DropMenu } from '../../shared/DropMenu'
 import UISettings from '../Settings/UISettings'
+import { devCycleAvatarSettings, isDevCommand } from './dev-util'
+import ChatOptions, { ChatModal } from './ChatOptions'
 
 const EDITING_KEY = 'chat-detail-settings'
-
-type Modal = 'export' | 'settings' | 'invite' | 'memory' | 'gen' | 'ui'
 
 const ChatDetail: Component = () => {
   const user = userStore()
@@ -65,7 +57,7 @@ const ChatDetail: Component = () => {
   const [swipe, setSwipe] = createSignal(0)
   const [removeId, setRemoveId] = createSignal('')
   const [showOpts, setShowOpts] = createSignal(false)
-  const [modal, setModal] = createSignal<Modal>()
+  const [modal, setModal] = createSignal<ChatModal>()
   const [editing, setEditing] = createSignal(getEditingState().editing ?? false)
   const { id } = useParams()
   const nav = useNavigate()
@@ -76,7 +68,7 @@ const ChatDetail: Component = () => {
     saveEditingState(next)
   }
 
-  const showModal = (modal: Modal) => {
+  const showModal = (modal: ChatModal) => {
     console.log('modal', modal)
     setModal(modal)
     setShowOpts(false)
@@ -112,18 +104,20 @@ const ChatDetail: Component = () => {
   })
 
   const sendMessage = (message: string, onSuccess?: () => void) => {
-    if (isDevCommand(message)) {
-      switch (message) {
-        case '/devCycleAvatarSettings':
-          devCycleAvatarSettings(user)
-      }
-    } else {
+    if (!isDevCommand(message)) {
       setSwipe(0)
       msgStore.send(chats.chat?._id!, message, false, onSuccess)
+      return
+    }
+
+    switch (message) {
+      case '/devCycleAvatarSettings':
+        devCycleAvatarSettings(user)
+        return
     }
   }
 
-  const moreMessage = (message: string) => {
+  const moreMessage = () => {
     msgStore.continuation(chats.chat?._id!)
   }
 
@@ -137,6 +131,7 @@ const ChatDetail: Component = () => {
 
   const headerBg = createMemo(() => {
     const rgb = getRootRgb('bg-900')
+    user.ui.mode // This 'unused' ref is needed to ensure this memo re-evaluated with the mode changes
     const styles: JSX.CSSProperties = {
       background: `rgba(${rgb.r}, ${rgb.g}, ${rgb.b}, 0.7)`,
     }
@@ -151,7 +146,7 @@ const ChatDetail: Component = () => {
         </div>
       </Show>
       <Show when={chats.chat}>
-        <div class="flex h-full flex-col justify-between sm:py-2 max-w-3xl mx-auto">
+        <div class="mx-auto flex h-full max-w-3xl flex-col justify-between sm:py-2">
           <div class="flex h-8 items-center justify-between rounded-md" style={headerBg()}>
             <div class="flex cursor-pointer flex-row items-center justify-between gap-4 text-lg font-bold">
               <Show when={!cfg.fullscreen}>
@@ -233,7 +228,7 @@ const ChatDetail: Component = () => {
                   )}
                 </For>
                 <Show when={msgs.waiting}>
-                  <div class="flex justify-center mt-3">
+                  <div class="mt-3 flex justify-center">
                     <div class="dot-flashing bg-[var(--hl-700)]"></div>
                   </div>
                 </Show>
@@ -285,76 +280,6 @@ const ChatDetail: Component = () => {
 
 export default ChatDetail
 
-const ChatOptions: Component<{
-  show: (modal: Modal) => void
-  editing: boolean
-  toggleEditing: () => void
-}> = (props) => {
-  const chats = chatStore((s) => ({ ...s.active, lastId: s.lastChatId }))
-  const user = userStore()
-
-  return (
-    <div class="flex w-60 flex-col gap-2 p-2">
-      <Show when={chats.chat?.userId === user.user?._id}>
-        <Option onClick={props.toggleEditing}>
-          <div class="flex w-full items-center justify-between">
-            <div>Enable Chat Editing</div>
-            <Toggle
-              class="flex items-center"
-              fieldName="editChat"
-              value={props.editing}
-              onChange={props.toggleEditing}
-            />
-          </div>
-        </Option>
-
-        <Option
-          onClick={() => props.show('invite')}
-          class="flex justify-start gap-2 hover:bg-[var(--bg-700)]"
-        >
-          <MailPlus /> Invite User
-        </Option>
-
-        <Option
-          onClick={() => props.show('memory')}
-          class="flex justify-start gap-2 hover:bg-[var(--bg-700)]"
-        >
-          <Book />
-          Edit Chat Memory
-        </Option>
-
-        <Option
-          onClick={() => props.show('gen')}
-          class="flex justify-start gap-2 hover:bg-[var(--bg-700)]"
-        >
-          <Sliders /> Generation Settings
-        </Option>
-
-        <Option
-          onClick={() => props.show('settings')}
-          class="flex justify-start gap-2 hover:bg-[var(--bg-700)]"
-        >
-          <Settings /> Chat Settings
-        </Option>
-      </Show>
-
-      <Option
-        onClick={() => props.show('ui')}
-        class="flex justify-start gap-2 hover:bg-[var(--bg-700)]"
-      >
-        <Palette /> UI Settings
-      </Option>
-
-      <Option
-        onClick={() => props.show('export')}
-        class="flex justify-start gap-2 hover:bg-[var(--bg-700)]"
-      >
-        <Download /> Export Chat
-      </Option>
-    </div>
-  )
-}
-
 const InviteModal: Component<{ chatId: string; show: boolean; close: () => void }> = (props) => {
   let ref: any
 
@@ -388,23 +313,6 @@ const InviteModal: Component<{ chatId: string; show: boolean; close: () => void 
         />
       </form>
     </Modal>
-  )
-}
-
-const Option: Component<{
-  children: any
-  class?: string
-  onClick: () => void
-  close?: () => void
-}> = (props) => {
-  const onClick = () => {
-    props.onClick()
-    props.close?.()
-  }
-  return (
-    <Button schema="secondary" size="sm" onClick={onClick}>
-      {props.children}
-    </Button>
   )
 }
 
@@ -471,35 +379,4 @@ function getEditingState() {
   const prev = localStorage.getItem(EDITING_KEY) || '{}'
   const body = JSON.parse(prev) as DetailSettings
   return body
-}
-
-/* Magic strings for dev testing purposes. */
-const devCommands = ['/devCycleAvatarSettings'] as const
-
-type DevCommand = (typeof devCommands)[number]
-
-const isDevCommand = (str: string): str is DevCommand => devCommands.some((cmd) => cmd === str)
-
-const devCycleAvatarSettings = (user: UserStore.State) => {
-  const originalSettings = {
-    avatarCorners: user.ui.avatarCorners,
-    avatarSize: user.ui.avatarSize,
-  }
-  const testedCornerSettings: AvatarCornerRadius[] = ['md', 'circle']
-  const settingPermutations = testedCornerSettings.flatMap((avatarCorners) =>
-    AVATAR_SIZES.map((avatarSize) => ({ avatarCorners, avatarSize }))
-  )
-  const applyPermutations = ([perm, ...rest]: (typeof settingPermutations)) => {
-    if (perm === undefined) {
-      console.log('Done demonstrating avatar setting permutations, restoring original settings')
-      const { avatarCorners, avatarSize } = originalSettings
-      userStore.updateUI({ avatarCorners, avatarSize })
-    } else {
-      console.log(perm)
-      const { avatarCorners, avatarSize } = perm
-      userStore.updateUI({ avatarCorners, avatarSize })
-      setTimeout(() => applyPermutations(rest), 800)
-    }
-  }
-  applyPermutations(settingPermutations)
 }
