+++ web/tailwind.css
@@ -122,17 +122,16 @@
   --black-800: #d1d5db;
   --black-900: #ffffff;
 
-  --text-100: #000000;
-  --text-200: #111827;
-  --text-300: #1f2937;
-  --text-400: #374151;
-  --text-500: #4b5563;
-  --text-600: #6b7280;
-  --text-700: #9ca3af;
-  --text-800: #d1d5db;
-  --text-900: #ffffff;
-
-  --chat-italics: #a3a3a3;
+  /** Uses --dark-* values */
+  --text-100: #171717;
+  --text-200: #262626;
+  --text-300: #404040;
+  --text-400: #525252;
+  --text-500: #737373;
+  --text-600: #a3a3a3;
+  --text-700: #d4d4d4;
+  --text-800: #e5e5e5;
+  --text-900: #f5f5f5;
 
   --highlight: rgb(14 165 233);
 }
